requirements.txt
streamlit>=1.28.0
python-barcode>=0.15.1
Pillow>=10.0.0
pandas>=1.5.0
